backend/requirements.txt
aiofiles==23.2.1
aistudio-sdk==0.3.8
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
asgiref==3.9.2
asyncpg==0.29.0
bce-python-sdk==0.9.60
bcrypt==4.0.1
beautifulsoup4==4.14.3
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
chardet==7.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cobble==0.1.4
colorlog==6.10.1
cryptography==46.0.3
cssselect==1.3.0
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
Deprecated==1.3.1
distro==1.9.0
Django==5.2
ecdsa==0.19.1
et_xmlfile==2.0.0
fastapi==0.128.0
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.20.3
fsspec==2026.1.0
future==1.0.0
google-auth==2.47.0
google-genai==1.60.0
greenlet==3.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
imagesize==2.0.0
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.3
kombu==5.6.2
limits==5.6.0
lxml==6.0.2
mammoth==1.11.0
MarkupSafe==3.0.3
modelscope==1.34.0
mpmath==1.3.0
networkx==3.6.1
newspaper3k==0.2.8
nltk==3.9.2
numpy==2.4.1
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
opencv-contrib-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==26.0
paddleocr==3.4.0
paddlepaddle==3.3.0
paddlex==3.4.2
pandas==3.0.0
passlib==1.7.4
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==11.3.0
prettytable==3.17.0
prompt_toolkit==3.0.52
protobuf==7.34.0
psutil==7.2.2
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pypdfium2==5.3.0
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.0
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==1.0.2
PyYAML==6.0.2
razorpay==2.0.0
redis==7.3.0
regex==2026.1.15
reportlab==4.4.9
requests==2.32.5
requests-file==3.0.1
rsa==4.9.1
ruamel.yaml==0.19.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.3
setuptools==80.10.1
sgmllib3k==1.0.0
shapely==2.1.2
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.23
sqlparse==0.5.3
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tinysegmenter==0.3
tldextract==5.3.1
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.1
transformers==4.57.6
triton==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.6.3
uvicorn==0.40.0
vine==5.1.0
warcio==1.7.5
wcwidth==0.6.0
websockets==15.0.1
wrapt==2.0.1
xlrd==2.0.2
xlsxwriter==3.2.9
